+++ src/test/model_service_test.cpp
@@ -29,6 +29,7 @@
 
 #include "../dags/pipelinedefinition.hpp"
 #include "../execution_context.hpp"
+#include "../grpc_utils.hpp"
 #include "../model_service.hpp"
 #include "../modelinstanceunloadguard.hpp"
 #include "../model_version_policy.hpp"
